requirements.txt
requests>=2.31.0
bcrypt>=4.0.1
python-dotenv>=1.0.0